requirements.txt
django~=4.0
stravalib~=0.10
django-allauth~=0.47
